SecureDropBox/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.3
idna==3.10
mysqlclient==2.2.7
Naked==0.1.32
pillow==11.2.1
pycparser==2.22
pycryptodome==3.23.0
python-magic==0.4.27
python-magic-bin==0.4.14
PyYAML==6.0.2
requests==2.32.4
shellescape==3.8.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
